Renamed  files to match the rest of the  project structure
